v.1.3: Added basic authentication and authorization;

+++ src/main/java/ru/inodinln/social_network/services/CommentService.java
@@ -4,7 +4,9 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.inodinln.social_network.entities.Comment;
+import ru.inodinln.social_network.entities.User;
 import ru.inodinln.social_network.exceptions.businessException.NotFoundException;
+import ru.inodinln.social_network.exceptions.securityException.AuthorizationException;
 import ru.inodinln.social_network.repositories.CommentRepository;
 
 import java.time.LocalDateTime;
@@ -58,10 +60,16 @@ public class CommentService {
     }
 
     @Transactional
-    public Comment create(Long postId, Long authorId, Long parentCommentId, Integer level, String text) {
+    public Comment create(Long postId, Long authorId, Long parentCommentId, Integer level,
+                          String text, String currentEmail) {
+        User user = userService.getById(authorId);
+
+        if(!(user.getEmail().equals(currentEmail)))
+            throw new AuthorizationException("Forbidden action with current credentials");
+
         Comment newComment = new Comment();
         newComment.setPost(postService.getById(postId));
-        newComment.setAuthor(userService.getById(authorId));
+        newComment.setAuthor(user);
         newComment.setLevel(level);
         newComment.setText(text);
         if (parentCommentId != null)
@@ -77,17 +85,27 @@ public class CommentService {
     }
 
     @Transactional
-    public Comment update(Long commentId, String text) {
+    public Comment update(Long commentId, String text, String currentEmail) {
         Comment commentToBeUpdated = getById(commentId);
+
+        if(!(userService.getByEmail(currentEmail).getId().equals(commentToBeUpdated.getAuthor().getId())))
+            throw new AuthorizationException("Forbidden action with current credentials");
+
         commentToBeUpdated.setText(text);
         commentToBeUpdated.setTimestampOfUpdating(LocalDateTime.now());
         return save(commentToBeUpdated);
     }
 
     @Transactional
-    public void delete(Long commentId) {
+    public void delete(Long commentId, String currentEmail) {
         Comment comment = getById(commentId);
+        User user = userService.getByEmail(currentEmail);
+
+        if(!(user.getId().equals(comment.getAuthor().getId()))
+        || !(user.getRole().name().equals("ROLE_ADMIN")))
+            throw new AuthorizationException("Forbidden action with current credentials");
+
         commentRepository.delete(comment);
-        postService.decreaseCountOfComments(comment.getPost().getId());
+        postService.decreaseCountOfComments(comment.getPost().getId()); //prop
     }
 }
